Upgrade to PooledLPToken models

+++ contracts/interfaces/IBondingCurve.sol
@@ -5,10 +5,10 @@ interface IBondingCurve {
     /**
     * @dev Given a reserve token amount, calculates the amount of continuous tokens returned.
     */
-    function getContinuousMintReward(uint _reserveTokenAmount) external view returns (uint);
+    function getContinuousMintReward(uint256 _reserveTokenAmount) external view returns (uint256);
 
     /**
     * @dev Given a continuous token amount, calculates the amount of reserve tokens returned.
     */  
-    function getContinuousBurnRefund(uint _continuousTokenAmount) external view returns (uint);
+    function getContinuousBurnRefund(uint256 _continuousTokenAmount) external view returns (uint256);
 }
